requirements.txt
cryptography
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
pymysql
sqlalchemy
